requirements.txt
asgiref==3.3.4
cffi==1.14.5
cryptography==3.4.7
Django==3.2.3
django-cors-headers==3.7.0
django-environ==0.4.5
django-filter==2.4.0
djangorestframework==3.12.4
djangorestframework-simplejwt==4.7.1
future==0.18.2
Markdown==3.3.4
pycparser==2.20
PyJWT==2.1.0
pytz==2021.1
serializers==0.2.4
sqlparse==0.4.1
uuid==1.30
